IOS-3913 Fix review comments
