+++ BlockchainSdk/WalletManagers/Chia/Common/Chia+Int64.swift
@@ -9,6 +9,7 @@
 import Foundation
 
 extension UInt64 {
+    // Use in any Chia entities
     func chiaEncode() -> Data {
         let data = withUnsafeBytes(of: self) { Data($0) }
         let result = data.bytes.reversed().drop(while: { $0 == 0x00 })
